refactor(l5): move promotion-service from src/ into server/domain behind an L5 adapter (issue 864)
Moved path:
- src/promotion-service.ts -> server/domain/promotion-service.ts

src/ closure under server/main.ts: 39 -> 38.

The server/ version takes one options object instead of the legacy
positional tail, so it holds no environment read of its own:
`targetNamespace`, `reason`, `auth`, `dryRun`, `now` and `killSwitch`
are all fields, and `promoteEntry(pool, table, id, options)` is four
parameters. `killSwitch` is what the old code read from
OPENBRAIN_PROMOTION_KILL_SWITCH inline.

server/tools/promote-entry.ts and server/tools/promotion.ts are
repointed at the new path and pass `killSwitch` from a new optional
`promotionKillSwitch` field on MemoryToolDependencies, wired in
server/main.ts from `config.promotion.killSwitch` next to the
`ftsCorpusConfig` and `searchEmbeddingTimeoutMs` fields that already
follow this pattern. `server/config/src-readers.ts` already parsed the
variable, so no new reader was needed and server/config.ts is untouched.

src/promotion-service.ts is now the L5 adapter (45 lines): it keeps the
old positional signature for the five legacy callers, reads
process.env.OPENBRAIN_PROMOTION_KILL_SWITCH inside the function at call
time, and delegates. Its type imports come from server/types.ts. Its
arguments arrive as one typed tuple rather than seven declared
parameters: the pre-commit gate lints the src/ adapter too and there is
no src/ exemption for max-params, so the tuple is what lets the
seven-argument legacy call form survive without a disable comment. The
call sites are unchanged and still type-checked position by position.

M4 seams found by linting the copy at the target path first: the four
`any` source-row parameters became a named PromotionSourceRow shape, the
repeated `extracted_metadata ?? metadata` chains became one
`metadataValue` helper (that was the complexity finding on the
provenance builder), and the 148-line promoteEntry body was split into
readSourceRow, assertTargetWritable and buildProvenance.

Tests are unchanged and stay in src/ per M1 EXCEPTION, passing through
the adapter: src/rest-promotion.test.ts,
src/tools/__tests__/promote-entry.pg.test.ts,
src/tools/__tests__/promote-shared.test.ts.
